+++ src/lib/plugins/RendererPlugin.ts
@@ -1,6 +1,6 @@
 import p5 from 'p5';
 import { P5AsciifyGrid } from '../Grid';
-import { P5AsciifyFontManager } from '../FontManager';
+import { P5AsciifyFontManager } from '../managers/FontManager';
 import { FeatureAsciiRendererOptions } from '../renderers/types';
 import { P5AsciifyAbstractFeatureRenderer2D } from '../renderers/2d/feature';
 import { P5AsciifyRenderer2D } from '../renderers/2d';
