+++ src/lib/renderers/2d/AsciiRenderer2D.ts
@@ -1,6 +1,6 @@
 import p5 from 'p5';
 import { P5AsciifyGrid } from '../../Grid';
-import { P5AsciifyFontManager } from '../../FontManager';
+import { P5AsciifyFontManager } from '../../managers/FontManager';
 import { P5AsciifyRenderer } from '../AsciiRenderer';
 import { AsciiRendererOptions } from '../types';
