+++ src/lib/renderers/2d/feature/AbstractFeatureRenderer2D.ts
@@ -1,6 +1,6 @@
 import p5 from 'p5';
 import { P5AsciifyGrid } from '../../../Grid';
-import { P5AsciifyFontManager } from '../../../FontManager';
+import { P5AsciifyFontManager } from '../../../managers/FontManager';
 import { P5AsciifyRenderer2D } from '../AsciiRenderer2D';
 import { FeatureAsciiRendererOptions } from '../../types';
 import { P5AsciifyError } from '../../../AsciifyError';
