+++ src/lib/utils/export/JSONExporter.ts
@@ -1,7 +1,7 @@
 import p5 from 'p5';
-import { P5AsciifyGrid } from '../Grid';
-import { P5AsciifyFontManager } from '../FontManager';
-import { P5AsciifyRendererManager } from '../renderers/RendererManager';
+import { P5AsciifyGrid } from '../../Grid';
+import { P5AsciifyFontManager } from '../../managers/FontManager';
+import { P5AsciifyRendererManager } from '../../renderers/RendererManager';
 
 /**
  * Options for JSON export.
