--- src/lib/utils/fonts/index.ts
@@ -1,2 +1,0 @@
-
-export * from './TyprFontUtils';
