+++ src/lib/utils/index.ts
@@ -1,9 +1,6 @@
-// Re-export specific utilities
 export { compareVersions } from './utils';
 
-// Re-export entire modules
-export * from './SVGExporter';
-export * from './JSONExporter';
+export * from './export/SVGExporter';
+export * from './export/JSONExporter';
 
-// Re-export the fonts submodule
-export * from './fonts';
+export * from './fonts/TyprFontUtils';
